agricultura/requirements.txt
dvc==3.0.0
yolov5==6.2.3
opencv-python==4.9.0.80
pandas==2.0.0
pillow==9.5.0
protobuf==3.9.2
imageio==2.34.0
mahotas==1.4.14
more_itertools==10.2.0
PyYAML==6.0.1
rawpy==0.19.1
tifffile==2023.7.10
tqdm==4.66.2
dulwich>=0.21.5
fsspec>=2022.10.0
scikit-image==0.20.0rc7
scikit-learn==1.3.2
matplotlib==3.7.5
xmltodict==0.13.0
seaborn==0.13.2
pycocotools==2.0.7
numpy==1.24.4
glob2==0.7
piexif==1.1.3
markdown_include==0.8.1
watchdog==4.0.0
weasyprint==61.2
click==8.1.3
yq==3.3.0
beautifulsoup4==4.12.3
